Initial version of fail2ban-subnets.
fail2ban-subnets aims to provide a way to ban subnets of
IPs repeatingly banned by fail2ban for multiple offenses.
It thus uses the fail2ban logfiles and calculates the most
restricted subnet to be banned for these IPs. Using the
log file generated by fail2ban-subnets, and a new action.d
script, we can thus create a specific jail in fail2ban for
banning those subnets.

fail2ban-subnets is here to provide what's currently
impossible in fail2ban, even if there are issues that
are progressing on that side.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>