Add an excludeban option to exclude from report the subnets already banned.
This option allows to avoid the "already banned" messages
from fail2ban that we get when reporting an already banned
IP. It will still verify that the subnet currently banned is
at least as large as the one we computed to ban. If it's
not the case, fail2ban-subnets will remove the previous ban
from iptables before reporting the new one in the logs,
making sure we're always as secure as possible.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>